dynamic equality constraint
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99993


ghstack-source-id: 187295230

This diff adds support for dynamic equality constraints of the form dynamic_dim(x, 0) == dynamic_dim(y, 1). The process of constraint discovery can already understand equality guards between dimensions and suggests such equality constraints, so this closes the loop on that. Correspondingly we now raise ConstraintViolation when we find that such a guard is added on a dynamic dimension and the user did not specify such a constraint. (NOTE: This is distinct from a dynamic dimension being guarded equal to a constant, which is already an error.)

Differential Revision: [D45279437](https://our.internmc.facebook.com/intern/diff/D45279437/)