[inductor] Share reduction combine_fn between backends
Currently reductions are hard-coded separately in `ir.py`, `codegen/triton.py`
and `codegen/cpp.py`. This reduces some of that duplication by factoring out the
`combine_fn` definition used in `_unroll_reduction_fn` and using it to generate
some of the reduction code in for the triton and cpp backends.

ghstack-source-id: d2377cda297d536ffb3848073da655df2302f59a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99661